requirements.txt
cloudscraper==1.2.71
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0

python-dotenv==1.0.1
Requests==2.32.3

SQLAlchemy~=2.0.36
alembic~=1.14.0